+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Initial support for BOOSTABLE CHARGES.  Associated burnout is not implemented.  Does not account for reducing the DC increase for powers with advantages. [#432](https://github.com/dmdorman/hero6e-foundryvtt/issues/432)
 - Fix for Combat Skill Levels where edit sheet did not allow for changing values.
 - Improved range pentaly tags and associated tooltips.
+- Fixed error for cone placement.
+
 
 # Version 3.0.50
 - Fix for 6e HDC import where some 5e values were incorrectly being used. [#430](https://github.com/dmdorman/hero6e-foundryvtt/issues/430)